+++ src/app/app.component.ts
@@ -1,4 +1,5 @@
 import {Component} from '@angular/core';
+import {Subject} from 'rxjs';
 import {TranslateService} from '@ngx-translate/core';
 import {NgbModal, NgbOffcanvas} from '@ng-bootstrap/ng-bootstrap';
 
@@ -15,9 +16,13 @@ import {DEFAULT_ACTIVE_MENU_ITEM_ID, MENU_ITEMS, MENU_ITEMS_IDS} from './app.con
     styleUrls: ['./app.component.scss']
 })
 export class AppComponent {
+    private _showScore$ = new Subject();
+    private showScore$$ = this._showScore$.asObservable();
+
     protected readonly MENU_ITEMS: Array<MenuItem> = MENU_ITEMS;
     protected readonly MENU_ITEMS_IDS = MENU_ITEMS_IDS;
     protected activeMenuItemId: string = DEFAULT_ACTIVE_MENU_ITEM_ID;
+    protected showScore: boolean = true;
 
     constructor(
         private atlasFilter: AtlasFilter,
@@ -31,6 +36,10 @@ export class AppComponent {
         translate.use('en-US');
 
         backendService.lifeIndexSubscription(this.atlasFilter);
+
+        this.showScore$$.subscribe(showScore => {
+            this.showScore = showScore as boolean;
+        });
     }
 
     onMenuItemClick(event: Event, menuItem: MenuItem): void {
@@ -55,6 +64,7 @@ export class AppComponent {
         const offcanvasRef = this.offcanvasService.open(SidebarComponent);
         offcanvasRef.componentInstance.name = 'Filter';
         offcanvasRef.componentInstance.onActiveButtonResets = this.onActiveButtonReset;
+        offcanvasRef.componentInstance.onToggleScore = this.onToggleScore;
         offcanvasRef.hidden.subscribe(value => {
             this.onActiveButtonReset();
         });
@@ -66,5 +76,10 @@ export class AppComponent {
 
     onActiveButtonReset(): void {
         this.activeMenuItemId = DEFAULT_ACTIVE_MENU_ITEM_ID;
+    }
+
+    onToggleScore = (showScore?: boolean) => {
+        const newShowScore = showScore || !this.showScore;
+        this._showScore$.next(newShowScore);
     }
 }