+++ src/app/views/atlas/atlas.types.ts
@@ -1,0 +1,7 @@
+import {GeoJSON, Layer} from 'leaflet';
+import {GeoFeature} from './constants/geo.types';
+
+export interface IAtlasLayer {
+    value: Layer | GeoJSON;
+    geoLand?: GeoFeature;
+}